fip: c1: update default amlogic ddw fw into T32 [1/1]
PD#SWPL-13619

Problem:
C1 BL2 needs to switch into T32, but default aml_ddr.fw was compiled
with A64, which would not work with T32 bl2.

Solution:
Recompile aml_ddr.fw with T32 and push into c1 folder for default.
So if we want to run C1 bl2 with A64, we need to manually update
ddw fw with A64, this is the expected option when we make C1 T32
bl2 as default.

Verify:
test pass on c1_ae409 boards.

Change-Id: I43a1b34ac103086b754a9a28eff38e43061f3265
Signed-off-by: honglin.zhang <honglin.zhang@amlogic.com>